# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ISP Ops Platform API
  version: 0.1.0
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
paths:
...
  /ipam/prefixes:
    get:
      summary: List prefixes
      responses:
        "200":
          description: Prefix list
    post:
      summary: Create prefix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prefix, family, role]
              properties:
                prefix:
                  type: string
                family:
                  type: integer
                  enum: [4, 6]
                role:
                  type: string
                siteCode:
                  type: string
                vrf:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Prefix created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
  /ipam/addresses:
    get:
      summary: List IP assignments with device, interface, site and service context
      responses:
        "200":
          description: IP assignment list
    post:
      summary: Create IP address assignment with optional interface context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, prefix, role]
              properties:
                address:
                  type: string
                prefix:
                  type: string
                role:
                  type: string
                interfaceId:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: IP address created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required and prefix must exist
...
  /monitoring/alerts:
    get:
      summary: List normalized alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Alert list
    post:
      summary: Create normalized alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectType, objectId, monitorSource, severity, title]
              properties:
                objectType:
                  type: string
                objectId:
                  type: string
                monitorSource:
                  type: string
                externalRef:
                  type: string
                severity:
                  type: string
                  enum: [critical, major, minor]
                title:
                  type: string
                description:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Alert created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
  /monitoring/alerts/{id}/ack:
    post:
      summary: Acknowledge alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert acknowledged
        "503":
          description: PostgreSQL required
  /monitoring/maintenance-windows:
    get:
      summary: List maintenance windows
      responses:
        "200":
          description: Maintenance window list
    post:
      summary: Create maintenance window
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectType, objectId, title, startsAt, endsAt]
              properties:
                objectType:
                  type: string
                objectId:
                  type: string
                title:
                  type: string
                startsAt:
                  type: string
                endsAt:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Maintenance window created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
...